+++ src/Application/Order/Invoice/ParsedInvoiceReferenceByYear.php
@@ -9,6 +9,11 @@ class ParsedInvoiceReferenceByYear
 {
     public function __construct(public readonly string $year, public readonly string $number)
     {
+        $number =  (int)$number;
+
+        if($number < 0) {
+            throw new \InvalidArgumentException('Invalid invoice number for invoice reference ' . $this->year . $this->number . '. Invoice number should be a positive integer.');
+        }
     }
 
     public static function fromInvoiceReference(InvoiceReference $invoiceReference): static
@@ -16,11 +21,11 @@ class ParsedInvoiceReferenceByYear
         $year = substr($invoiceReference->get(), 0, 2);
         $number = substr($invoiceReference->get(), 2);
 
-        return new static($year, $number);
+        return new self($year, $number);
     }
 
     public function getNumberAsInt(): int
     {
-        return (int)$this->number;
+        return (int) $this->number;
     }
 }